backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SocketIO==5.6.1
bcrypt==4.2.0
python-dotenv==1.0.1
openpyxl==3.1.5
